// src/types/types.ts
// Types for dark mode
export type darkModeType = {
  isDarkMode: boolean;
};
//Types for github user
export type githubUser = {
  name: string;
  username: string;
  company: string;
  location: string;
  bio: string;
  twitter_username: string;
  public_repos: number;
  followers: number;
  following: number;
  profile_url: string;
  avatar_url: string;
  created_at: string;
};